added authentication

+++ app/api/issues/[id]/route.ts
@@ -5,28 +5,36 @@ import { issueSchema } from "@/app/validationSchemas";
 import prisma from "@/prisma/client";
 import { NextRequest, NextResponse } from "next/server";
 import delay from 'delay'
+import authOptions from "@/app/auth/authOptions";
+import { request } from "http";
+import { getServerSession } from "next-auth";
+import { string } from "zod";
 
 export async function PATCH(
     request: NextRequest,
-    {params}: {params: {id: string}}){
+    { params }: { params: { id: string } }) {
+
+    const session = await getServerSession(authOptions);
+    if (!session)
+        return NextResponse.json({}, { status: 401 });
 
     const body = await request.json();
     const validation = issueSchema.safeParse(body);
 
     if (!validation.success)
-        return NextResponse.json(validation.error.format(), {status: 400});
+        return NextResponse.json(validation.error.format(), { status: 400 });
 
 
     const issue = await prisma.issue.findUnique({
-        where: {id: parseInt(params.id)}
+        where: { id: parseInt(params.id) }
     });
 
     if (!issue)
-        return NextResponse.json({error: 'Invalid Isse'}, {status: 404});
-    
+        return NextResponse.json({ error: 'Invalid Isse' }, { status: 404 });
+
 
     const updatedIssue = await prisma.issue.update({
-        where: { id: issue.id},
+        where: { id: issue.id },
         data: {
             title: body.title,
             descrition: body.description
@@ -39,19 +47,21 @@ export async function PATCH(
 
 export async function DELETE(
     request: NextRequest,
-    {params}: {params: {id: string}}){
+    { params }: { params: { id: string } }) {
 
+    const session = await getServerSession(authOptions);
+    if (!session)
+        return NextResponse.json({}, { status: 401 });
 
-   
     const issue = await prisma.issue.findUnique({
-        where: {id: parseInt(params.id)}
+        where: { id: parseInt(params.id) }
     });
 
     if (!issue)
-        return NextResponse.json({error: 'Invalid Issue'}, {status: 404});
+        return NextResponse.json({ error: 'Invalid Issue' }, { status: 404 });
 
     await prisma.issue.delete({
-        where: {id: issue.id}
+        where: { id: issue.id }
     });
 
     return NextResponse.json({});
